Implemented money generation based on number of greggs

+++ core/src/main/java/com/spacecomplexity/longboilife/game/ui/game/UIMoneyMenu.java
@@ -47,7 +47,7 @@ public class UIMoneyMenu extends UIElement {
     }
 
     public void render() {
-        label.setText(NumberFormat.getCurrencyInstance(Locale.UK).format(GameState.getState().money));
+        label.setText(NumberFormat.getCurrencyInstance(Locale.UK).format((int) GameState.getState().money));
     }
 
     @Override
